+++ webui/cypress/integration/filtering.spec.js
@@ -65,19 +65,19 @@ describe('Replicaset filtering', () => {
     cy.contains('Replicasets');
 
     // Healthy
-    cy.get('input[placeholder="Filter by uri, uuid, role, alias "]').prev('button').click();
+    cy.get('button[type="button"]:contains(Filter)').click();
     cy.get('.meta-test__Filter__Dropdown *:contains(Healthy)').click({ force: true });
     cy.get('.meta-test__Filter input').should('have.value', 'status:healthy');
     cy.get('.ServerLabelsHighlightingArea').contains('test-storage-1').should('not.exist');
 
     // Unhealthy
-    cy.get('input[placeholder="Filter by uri, uuid, role, alias "]').prev('button').click();
+    cy.get('button[type="button"]:contains(Filter)').click();
     cy.get('.meta-test__Filter__Dropdown *:contains(Unhealthy)').click({ force: true });
     cy.get('.meta-test__Filter input').should('have.value', 'status:unhealthy');
     cy.get('.ServerLabelsHighlightingArea').contains('test-storage-1');
 
     // Role
-    cy.get('input[placeholder="Filter by uri, uuid, role, alias "]').prev('button').click();
+    cy.get('button[type="button"]:contains(Filter)').click();
     cy.get('.meta-test__Filter__Dropdown').find('*:contains(vshard-storage)').click({ force: true });
     cy.get('.meta-test__Filter input').should('have.value', 'role:vshard-storage');
     cy.get('.ServerLabelsHighlightingArea').contains('test-storage-1');
@@ -88,7 +88,7 @@ describe('Replicaset filtering', () => {
     cy.get('.meta-test__Filter svg').eq(1).click();
 
     // Search
-    cy.get('input[placeholder="Filter by uri, uuid, role, alias "]').type('test-storage-1');
+    cy.get('.meta-test__Filter').find('input').type('test-storage-1');
     cy.get('#root').contains('test-storage');
     cy.get('#root').contains('test-router').should('not.exist');