+++ webui/cypress/integration/webui-features.spec.js
@@ -146,15 +146,13 @@ describe('Web UI features testing', () => {
 
     for (let i = 1; i <= 7; i++) {
       it(`Check ${filterValue.get(i)} ${filteredValue.get(i)} `, () => {
-        cy.get('input[placeholder="Filter by uri, uuid, role, alias "]').prev('button').click();
+        cy.get('button[type="button"]:contains(Filter)').click();
         cy.get('.meta-test__Filter__Dropdown *')
           .contains(new RegExp('^' + filterValue.get(i) + '$'))
           .click({ force: true });
-        let value = filteredValue.get(i);
-        cy.get('input[placeholder="Filter by uri, uuid, role, alias "]').should('have.value', `${value}`);
+        cy.get('.meta-test__Filter input').should('have.value', `${filteredValue.get(i)}`);
         cy.reload(true);
-        value = filteredValue.get(i);
-        cy.get('input[placeholder="Filter by uri, uuid, role, alias "]').should('have.value', `${value}`);
+        cy.get('.meta-test__Filter input').should('have.value', `${filteredValue.get(i)}`);
       });
     }
   });