+++ webui/src/models/cluster/server-list/gate.ts
@@ -1,8 +1,0 @@
-import { createGate } from 'effector-react';
-
-export type FilterGateType = {
-  rolesFilter: string;
-  ratingFilter: string;
-};
-
-export const FilterGate = createGate<FilterGateType>({});