+++ webui/src/pages/Cluster/components/ReplicasetFilter/index.ts
@@ -1,2 +1,0 @@
-export { default } from './ReplicasetFilter';
-export type { ReplicasetFilterType } from './ReplicasetFilter';