+++ webui/src/pages/Cluster/components/ReplicasetFilterInput/ReplicasetFilterInput.styles.ts
@@ -3,10 +3,10 @@ import { css } from '@emotion/css';
 export const styles = {
   root: css`
     position: relative;
+
     input {
       width: 100%;
     }
-    margin-bottom: 10px;
   `,
   chevron: css`
     fill: rgba(245, 34, 45, 0.65);