+++ webui/src/pages/Cluster/components/ReplicasetListPageSection/ReplicasetListPageSection.tsx
@@ -1,6 +1,7 @@
 /* eslint-disable @typescript-eslint/ban-ts-comment */
-import React, { useMemo } from 'react';
+import React, { useEffect, useMemo, useState } from 'react';
 import { useStore } from 'effector-react';
+import { useCore } from '@tarantool.io/frontend-core';
 // @ts-ignore
 import { PageSection } from '@tarantool.io/ui-kit';
 
@@ -11,13 +12,18 @@ import ReplicasetListHeader from './components/ReplicasetListHeader';
 
 import { styles } from './ReplicasetListPageSection.styles';
 
-const { $serverList, $cluster, $filteredReplicasetList, selectors } = models.cluster.serverList;
+const { $serverList, $cluster, selectors, filters } = models.cluster.serverList;
 
 const ReplicasetListPageSection = () => {
+  const core = useCore();
   const serverListStore = useStore($serverList);
   const clusterStore = useStore($cluster);
 
-  const filteredSearchableReplicasetList = useStore($filteredReplicasetList);
+  const [filter, setFilter] = useState(core?.ss.get('cluster_filter') || '');
+
+  useEffect(() => {
+    core?.ss.set('cluster_filter', filter);
+  }, [core, filter]);
 
   const [issues, replicasetList, serverStat] = useMemo(
     () => [
@@ -37,14 +43,21 @@ const ReplicasetListPageSection = () => {
     [clusterStore]
   );
 
+  const replicasetListSearchable = useMemo(() => selectors.replicasetListSearchable(replicasetList), [replicasetList]);
+
+  const filteredSearchableReplicasetList = useMemo(
+    () => filters.filterSearchableReplicasetList(replicasetListSearchable, filter),
+    [replicasetListSearchable, filter]
+  );
+
   if (replicasetList.length < 1 || !cluster || !clusterSelf) {
     return null;
   }
 
   return (
     <PageSection title="Replicasets">
-      <ReplicasetListHeader />
-      {filteredSearchableReplicasetList?.length ? (
+      <ReplicasetListHeader filter={filter} onFilterChange={setFilter} />
+      {filteredSearchableReplicasetList.length ? (
         <ReplicasetList
           cluster={cluster}
           clusterSelf={clusterSelf}